Add files via upload
A smart attendance application harnesses facial recognition and geofencing technology to streamline attendance tracking. This innovative solution automates the traditional attendance process by capturing a student or employee's face upon entry and cross-referencing it with a database of authorized individuals. Geofencing ensures that attendance is recorded only when the person is within a predefined location, enhancing accuracy. This system not only saves time but also enhances security by preventing proxy attendance. This combination of facial recognition and geofencing technology brings efficiency, accuracy, and security to attendance management in various educational and organizations.